+++ libraries/base/changelog.md
@@ -1,6 +1,10 @@
 # Changelog for [`base` package](http://hackage.haskell.org/package/base)
 
 
+## 4.11.1.0 *TBA*
+  * Add the `readFieldHash` function to `GHC.Read` which behaves like
+    `readField`, but for a field that ends with a `#` symbol (#14918).
+
 ## 4.11.0.0 *TBA*
   * Bundled with GHC 8.4.1